인증 모듈 사용

+++ controller/OrderController.js
@@ -1,4 +1,5 @@
-// const conn = require("../mariadb"); // db 모듈
+const ensureAuthorization = require("../auth"); // 인증 모듈
+const jwt = require("jsonwebtoken");
 const mariadb = require("mysql2/promise");
 const { StatusCodes } = require("http-status-codes");
 
@@ -11,39 +12,56 @@ const order = async (req, res) => {
     dateStrings: true,
   });
 
-  const { items, delivery, totalQuantity, totalPrice, userId, firstBookTitle } =
+  const { items, delivery, totalQuantity, totalPrice, firstBookTitle } =
     req.body;
+  const authorization = ensureAuthorization(req, res);
 
-  // delivery 테이블 삽입
-  let sql = `INSERT INTO delivery(address, receiver, contact) VALUES (?, ?, ?);`;
-  let values = [delivery.address, delivery.receiver, delivery.contact];
-  let [results] = await conn.execute(sql, values);
-  let delivery_id = results.insertId;
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res
+      .status(StatusCodes.UNAUTHORIZED)
+      .json({ message: "로그인 세션이 만료되었습니다. 다시 로그인 해주세요." });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res
+      .status(StatusCodes.BAD_REQUEST)
+      .json({ message: "잘못된 토큰입니다." });
+  } else {
+    // delivery 테이블 삽입
+    let sql = `INSERT INTO delivery(address, receiver, contact) VALUES (?, ?, ?);`;
+    let values = [delivery.address, delivery.receiver, delivery.contact];
+    let [results] = await conn.execute(sql, values);
+    let delivery_id = results.insertId;
 
-  // orders 테이블 삽입
-  sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id)
-         VALUES (?, ?, ?, ?, ?);`;
-  values = [firstBookTitle, totalQuantity, totalPrice, userId, delivery_id];
-  [results] = await conn.execute(sql, values);
-  let order_id = results.insertId;
+    // orders 테이블 삽입
+    sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id)
+       VALUES (?, ?, ?, ?, ?);`;
+    values = [
+      firstBookTitle,
+      totalQuantity,
+      totalPrice,
+      authorization.id,
+      delivery_id,
+    ];
+    [results] = await conn.execute(sql, values);
+    let order_id = results.insertId;
 
-  // items를 가지고, 장바구니에서 book_id, quantity 조회
-  sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?);`;
-  let [orderItems, fields] = await conn.query(sql, [items]);
+    // items를 가지고, 장바구니에서 book_id, quantity 조회
+    sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?);`;
+    let [orderItems, fields] = await conn.query(sql, [items]);
 
-  // orderedBook 테이블 삽입
-  sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
+    // orderedBook 테이블 삽입
+    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
 
-  // items.. 배열: 요소들을 하나씩 꺼내서 (foreach문 돌려서) >
-  values = [];
-  orderItems.forEach((item) => {
-    values.push([order_id, item.book_id, item.quantity]);
-  });
-  results = await conn.query(sql, [values]);
+    // items.. 배열: 요소들을 하나씩 꺼내서 (foreach문 돌려서) >
+    values = [];
+    orderItems.forEach((item) => {
+      values.push([order_id, item.book_id, item.quantity]);
+    });
+    results = await conn.query(sql, [values]);
 
-  let result = await deleteCartItems(conn, items);
+    let result = await deleteCartItems(conn, items);
 
-  return res.status(StatusCodes.OK).json(result);
+    return res.status(StatusCodes.OK).json(result);
+  }
 };
 
 const deleteCartItems = async (conn, items) => {
@@ -72,7 +90,7 @@ const getOrders = async (req, res) => {
 };
 
 const getOrderDetail = async (req, res) => {
-  const { id } = req.params;
+  const orderId = req.params.id;
   const conn = await mariadb.createConnection({
     host: "127.0.0.1",
     user: "root",
@@ -85,7 +103,7 @@ const getOrderDetail = async (req, res) => {
              FROM orderedBook LEFT JOIN books
              ON orderedBook.book_id = books.id
              WHERE order_id = ?;`;
-  const [rows, fields] = await conn.query(sql, [id]);
+  const [rows, fields] = await conn.query(sql, [orderId]);
   return res.status(StatusCodes.OK).json(rows);
 };
